Fixed message timing and messages after death
The message saying the player died would appear 0.5 seconds before the
actual death. It now appears 0.1 seconds before (not 0 to avoid
IsPlayerAlive checks).

Made it so if the player is dead when the kill message is displayed, it
will instead print "[SM] You have to be alive to use RocketMe!" to the
client.